refactor: rename /integrations to /channels
- Rename routes/integrations/ folder to routes/channels/
- Update basePath in telegram.ts and discord.ts
- Update imports in index.ts
- Update all frontend API calls to use /channels/ endpoints
- Update CLAUDE.md documentation
