+++ apps/api/src/index.ts
@@ -8,8 +8,8 @@ import adminStats from "./routes/admin/stats";
 import adminUsers from "./routes/admin/users";
 import dashboard from "./routes/dashboard";
 import health from "./routes/health";
-import discord from "./routes/integrations/discord";
-import telegram from "./routes/integrations/telegram";
+import discord from "./routes/channels/discord";
+import telegram from "./routes/channels/telegram";
 import repos from "./routes/repos";
 import stats from "./routes/stats";
 import webhook from "./routes/webhook";
